requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.7
cx-Oracle==8.3.0
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.59.0
django-crispy-forms==2.1
django-crontab==0.7.1
django-environ==0.11.2
django-resized==1.0.2
django-session-timeout==0.1.0
idna==3.6
mysqlclient==2.2.1
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.0
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
